@@ -12,23 +12,23 @@
 		"format": "prettier --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
-		"@material/web": "1.0.0-pre.11",
-		"@sveltejs/kit": "^2.5.17",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
-		"eslint": "^8.57.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.41.0",
-		"firebase": "^10.12.2",
-		"mdsvex": "^0.11.2",
-		"prettier": "^3.3.2",
-		"prettier-plugin-svelte": "^3.2.5",
-		"sass": "^1.77.6",
-		"svelte": "^4.2.18",
-		"vite": "^5.3.1"
+		"@material/web": "2.4.1",
+		"@sveltejs/kit": "^2.51.0",
+		"@sveltejs/vite-plugin-svelte": "^6.2.4",
+		"eslint": "^10.0.0",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.15.0",
+		"firebase": "^12.9.0",
+		"mdsvex": "^0.12.6",
+		"prettier": "^3.8.1",
+		"prettier-plugin-svelte": "^3.4.1",
+		"sass": "^1.97.3",
+		"svelte": "^5.0.0",
+		"vite": "^7.3.1"
 	},
 	"dependencies": {
-		"@sveltejs/adapter-vercel": "^5.3.2",
-		"fast-xml-parser": "^4.4.0"
+		"@sveltejs/adapter-vercel": "^6.3.1",
+		"fast-xml-parser": "^5.3.6"
 	},
-	"packageManager": "pnpm@9.4.0+sha512.f549b8a52c9d2b8536762f99c0722205efc5af913e77835dbccc3b0b0b2ca9e7dc8022b78062c17291c48e88749c70ce88eb5a74f1fa8c4bf5e18bb46c8bd83a"
+	"packageManager": "pnpm@10.29.3"
 }
