@@ -2,8 +2,7 @@
     import "@material/web/icon/icon.js";
     import Card from "../../../components/Card.svelte";
 
-    /** @type {import('./$types').PageData} */
-    export let data;
+    let { data } = $props();
 
 
 </script>
