@@ -1,8 +1,7 @@
 <script>
     import Card from '../../../components/Card.svelte';
 
-    /** @type {import('./$types').PageData} */
-    export let data;
+    let { data } = $props();
 </script>
 
 <svelte:head>
